refactor(TierPage): move tier page query to the queries file
